backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.4
cachetools==5.5.0
google-generativeai==0.8.2
pydantic==2.9.2
python-dotenv==1.0.1
numpy
gunicorn
